+++ package.json
@@ -44,6 +44,7 @@
     "utility-types": "^3.10.0"
   },
   "devDependencies": {
+    "@ant-design/bisheng-plugin": "^2.3.0",
     "@babel/core": "^7.6.2",
     "@babel/runtime": "^7.6.2",
     "@react-native-community/cameraroll": "^1.5.2",
@@ -63,33 +64,36 @@
     "ant-design-palettes": "^1.1.3",
     "antd": "3.11.1",
     "antd-mobile-demo-data": "^0.2.0",
-    "antd-tools": "^5.2.6",
+    "antd-tools": "^7.7.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^7.2.3",
     "babel-jest": "^25.4.0",
     "babel-plugin-import": "^1.13.0",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-react-native": "^4.0.1",
-    "bisheng": "^0.28.0",
-    "bisheng-plugin-antd": "^0.17.0",
-    "bisheng-plugin-description": "^0.1.1",
-    "bisheng-plugin-react": "^0.6.0",
-    "bisheng-plugin-toc": "^0.4.0",
+    "bisheng": "^1.5.3",
+    "bisheng-plugin-description": "^0.1.4",
+    "bisheng-plugin-react": "^1.1.0",
+    "bisheng-plugin-toc": "^0.4.4",
     "classnames": "^2.2.1",
     "codesandbox": "^1.3.5",
     "concurrently": "^4.0.1",
     "cross-env": "^5.0.1",
-    "css-split-webpack-plugin": "^0.2.5",
+    "css-split-webpack-plugin": "^0.2.6",
     "dora-plugin-upload": "^0.3.1",
     "enquire.js": "^2.1.1",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "~15.1.0",
-    "eslint-plugin-babel": "~4.1.1",
-    "eslint-plugin-import": "~2.7.0",
-    "eslint-plugin-jsx-a11y": "~6.0.2",
-    "eslint-plugin-markdown": "~1.0.0-beta.6",
-    "eslint-plugin-react": "~7.2.0",
-    "eslint-tinker": "~0.4.1",
+    "eslint": "^7.0.0",
+    "eslint-config-airbnb": "^18.0.0",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.21.1",
+    "eslint-plugin-jest": "^23.0.2",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-markdown": "^1.0.0",
+    "eslint-plugin-react": "^7.14.2",
+    "eslint-plugin-react-hooks": "^4.0.0",
+    "eslint-plugin-unicorn": "^20.0.0",
+    "eslint-tinker": "^0.5.0",
     "glob": "^7.1.1",
     "jest": "^25.4.0",
     "jsonml.js": "^0.1.0",
@@ -138,8 +142,8 @@
     "compile": "cross-env BABEL_ENV=dist antd-tools run compile",
     "watch-tsc": "antd-tools run watch-tsc",
     "clean": "antd-tools run clean",
-    "start": "cross-env NODE_ENV=development cross-env DEMO_ENV=react bisheng start -c ./site/bisheng.desktop.config.js",
-    "site": "cross-env NODE_ENV=production cross-env DEMO_ENV=react bisheng build -c ./site/bisheng.desktop.config.js && cp CNAME _site",
+    "start": "cross-env NODE_ENV=development cross-env DEMO_ENV=react bisheng start -c ./site/bisheng.config.js",
+    "site": "cross-env NODE_ENV=production cross-env DEMO_ENV=react bisheng build -c ./site/bisheng.config.js && cp CNAME _site",
     "deploy": "npm run clean && npm run site && bisheng gh-pages --push-only",
     "pub": "cross-env BABEL_ENV=dist antd-tools run pub",
     "rn-start": "node scripts/start.js",
