--- site/bisheng.desktop.config.js
@@ -1,16 +1,0 @@
-const path = require('path');
-const commonConfig = require('./bisheng.common.config');
-
-module.exports = Object.assign({}, commonConfig, {
-  port: 8003,
-  source: {
-    components: './components',
-    docs: './docs',
-    changelog: [
-      'CHANGELOG.zh-CN.md',
-      'CHANGELOG.en-US.md',
-    ],
-  },
-  theme: './site/desktop/src',
-  htmlTemplate: path.join(__dirname, './desktop/src/static/template.html'),
-});
