--- site/desktop/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "bisheng-theme-antd-mobile-desktop",
-  "version": "0.1.0",
-  "main": "lib/index.js",
-  "file": [
-    "lib"
-  ],
-  "scripts": {
-    "prepublish": "babel src --out-dir lib && cp -r src/static/* lib/static",
-    "postpublish": "rm -rf lib"
-  },
-  "babel": {
-    "presets": [
-      "react",
-      [
-        "env",
-        {
-          "targets": {
-            "browsers": [
-              "last 2 versions",
-              "Firefox ESR",
-              "> 1%",
-              "ie >= 8",
-              "iOS >= 8",
-              "Android >= 4"
-            ]
-          }
-        }
-      ]
-    ],
-    "plugins": [
-      "transform-class-properties",
-      "transform-object-rest-spread"
-    ]
-  },
-  "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "less": "^2.7.2"
-  },
-  "license": "MIT"
-}
