+++ site/theme/index.js
@@ -53,7 +53,7 @@ module.exports = {
   plugins: [
     'bisheng-plugin-description',
     'bisheng-plugin-toc?maxDepth=2&keepElem',
-    'bisheng-plugin-antd?noPreview',
+    '@ant-design/bisheng-plugin?injectProvider',
     'bisheng-plugin-react?lang=__react',
   ],
   routes: {
