+++ site/theme/static/new-version-info-modal.less
