+++ site/theme/static/not-found.less
